packages/iso8583-core/pyproject.toml
[project]
name = "iso8583-core"
version = "0.1.0"
description = "ISO 8583 ユースケース・インフラ実装"
requires-python = ">=3.10"
dependencies = [
    "iso8583-types",
    "pyiso8583>=1.0.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.uv.sources]
iso8583-types = { path = "../iso8583-types", editable = true }
